Add files via upload
A recommender system is a system which gives recommendation of items such as movies, music, books, news, images, web pages, tools to a user. This information is processed so that it is more close to interest of users. The aim of a recommender system is often to help users learn about new products ones among variety of choices.
It aims at recommending relevant information and removing unwanted or redundant information